Update dataconnection.js
After completing a large file transfer, it seems like the `_chunkedData` object is not being cleared frequently enough, and so it accumulates data and eats up memory. I think the issue is due to the fact that `delete this._chunkedData[id];` is executed after the recursive call to `_handleDataMessage()`.

To fix, execute `delete this._chunkedData[id];` before the recursive call to `_handleDataMessage()` to free up memory and prevent `_chunkedData` from accumulating a bunch of data.